Moved CacheMixin class to separate file mixin.py
